refactor: useFocusControlをMultiComboboxと統合
