+++ packages/smarthr-ui/src/components/Combobox/MultiCombobox/MultiCombobox.tsx
@@ -7,15 +7,16 @@ import {
   type MouseEvent,
   type ReactNode,
   type Ref,
+  createRef,
   memo,
   useCallback,
   useEffect,
+  useId,
   useImperativeHandle,
   useMemo,
   useRef,
   useState,
 } from 'react'
-import { useId } from 'react'
 import innerText from 'react-innertext'
 import { tv } from 'tailwind-variants'
 
@@ -31,7 +32,6 @@ import { ListBox, useListbox } from '../useListbox'
 import { useMultiOptions } from '../useOptions'
 
 import { MultiSelectedItem } from './MultiSelectedItem'
-import { useFocusControl } from './useFocusControl'
 
 import type { ComboboxItem, BaseProps as ComboboxProps } from '../types'
 
@@ -275,13 +275,61 @@ const ActualMultiCombobox = <T,>(
     noResultText,
   })
 
-  const {
-    deletionButtonRefs,
-    inputRef,
-    focusPrevDeletionButton,
-    focusNextDeletionButton,
-    resetDeletionButtonFocus,
-  } = useFocusControl(selectedItems.length)
+  const selectedItemLength = selectedItems.length
+
+  const deletionButtonRefs = useMemo(() => {
+    const refs: Array<ReturnType<typeof createRef<HTMLButtonElement>>> = []
+
+    for (let i = 0; i < selectedItemLength; i++) {
+      refs[i] = createRef<HTMLButtonElement>()
+    }
+
+    return refs
+  }, [selectedItemLength])
+  const inputRef = useRef<HTMLInputElement>(null)
+
+  const [focusedIndex, setFocusedIndex] = useState<number | null>(null)
+
+  const focusPrevDeletionButton = useCallback(() => {
+    if (selectedItemLength === 0) {
+      return
+    }
+
+    if (focusedIndex !== null) {
+      const nextIndex = Math.max(focusedIndex - 1, 0)
+
+      deletionButtonRefs[nextIndex].current?.focus()
+      setFocusedIndex(nextIndex)
+    } else if (inputRef.current?.selectionStart === 0) {
+      const nextIndex = deletionButtonRefs.length - 1
+
+      deletionButtonRefs[nextIndex].current?.focus()
+      setFocusedIndex(nextIndex)
+    }
+  }, [deletionButtonRefs, focusedIndex, selectedItemLength])
+
+  const focusNextDeletionButton = useCallback(() => {
+    if (deletionButtonRefs.length === 0 || focusedIndex === null) {
+      return
+    }
+
+    const nextIndex = focusedIndex + 1
+
+    if (nextIndex < deletionButtonRefs.length) {
+      deletionButtonRefs[nextIndex].current?.focus()
+      setFocusedIndex(nextIndex)
+    } else {
+      setFocusedIndex(null)
+      // キー入力が input に影響しないようにフォーカスタイミングを遅らせる
+      setTimeout(() => {
+        inputRef.current?.focus()
+      })
+    }
+  }, [deletionButtonRefs, focusedIndex])
+
+  const resetDeletionButtonFocus = useCallback(() => {
+    setFocusedIndex(null)
+  }, [])
 
   useImperativeHandle<HTMLInputElement | null, HTMLInputElement | null>(ref, () => inputRef.current)
